feat: renamed tabs and set collection the  as default
