+++ src/routes/(public)/user/[username]/+page.server.ts
@@ -1,44 +1,35 @@
 import type { PageServerLoad } from "./$types";
 import { logError } from "$lib/server/logError";
 
-export const load = (async ({ parent, locals: { log, supabase } }) => {
+export const load = (async ({ parent, locals: { supabase, log } }) => {
   const { profile } = await parent();
 
   const [
-    { data: user_cubes, error: ucErr },
-    { data: user_achievements, error: uaErr },
-    { data: user_cube_ratings, error: urErr },
+    { data: user_cubes, error: userCubesError },
+    { data: user_cube_ratings, error: userRatingsError },
   ] = await Promise.all([
     supabase
       .from("user_cubes")
-      .select(`*, cube_model:cube_models(*), vendor:bought_from(name)`)
-      .eq("user_id", profile.user_id),
-    supabase
-      .from("user_achievements")
-      .select("*, achievement:achievement_slug(*)")
+      .select(
+        "*, cube_model:v_detailed_cube_models(*), bought_from, vendor:bought_from(name)",
+      )
       .eq("user_id", profile.user_id),
     supabase
       .from("user_cube_ratings")
-      .select("*, cube_model:cube_models(*)")
+      .select("*")
       .eq("user_id", profile.user_id),
   ]);
 
-  if (ucErr) {
-    return logError(500, "Unable to load user cubes", log, ucErr);
+  if (userCubesError) {
+    return logError(500, "Unable to load user cubes", log, userCubesError);
   }
-  if (uaErr) {
-    return logError(500, "Unable to load user achievements", log, uaErr);
+  if (userRatingsError) {
+    return logError(500, "Unable to load user ratings", log, userRatingsError);
   }
-  if (urErr) {
-    return logError(500, "Unable to load user ratings", log, urErr);
-  }
-
-  const main_cubes = user_cubes.filter((uc) => uc.main === true);
 
   return {
-    main_cubes,
     user_cubes,
-    user_achievements,
     user_cube_ratings,
+    meta: { title: `${profile.display_name}'s Cube Collection - CubeIndex` },
   };
 }) satisfies PageServerLoad;
